starting to clean up the decoder/simulator
