+++ .gitignore
@@ -28,3 +28,4 @@ data/sdss/rcf_list
 data/sdss_all
 case_studies/sdss_galaxies/models/simulated_blended_galaxies.pt
 venv
+nohup.out
