--- bliss/simulator/galsim_galaxies.py
@@ -1,74 +1,0 @@
-from typing import Dict, List, Optional
-
-import pytorch_lightning as pl
-import torch
-from torch import Tensor
-from torch.utils.data import DataLoader, Dataset
-from tqdm import tqdm
-
-from bliss.simulator.background import ConstantBackground
-from bliss.simulator.galsim_decoder import SingleGalsimGalaxyDecoder, SingleGalsimGalaxyPrior
-
-
-class SingleGalsimGalaxies(pl.LightningDataModule, Dataset):
-    def __init__(
-        self,
-        prior: SingleGalsimGalaxyPrior,
-        decoder: SingleGalsimGalaxyDecoder,
-        background: ConstantBackground,
-        num_workers: int,
-        batch_size: int,
-        n_batches: int,
-        fix_validation_set: bool = False,
-        valid_n_batches: Optional[int] = None,
-    ):
-        super().__init__()
-        self.prior = prior
-        self.decoder = decoder
-        self.n_batches = n_batches
-        self.batch_size = batch_size
-        self.num_workers = num_workers
-        self.background = background
-        self.fix_validation_set = fix_validation_set
-        self.valid_n_batches = valid_n_batches
-
-    def __getitem__(self, idx):
-        galaxy_params = self.prior.sample(1)
-        galaxy_image = self.decoder(galaxy_params[0])
-        background = self.background.sample((1, *galaxy_image.shape)).squeeze(1)
-        return {
-            "images": _add_noise_and_background(galaxy_image, background),
-            "background": background,
-            "noiseless": galaxy_image,
-            "params": galaxy_params[0],
-            "snr": _get_snr(galaxy_image, background),
-        }
-
-    def __len__(self):
-        return self.batch_size * self.n_batches
-
-    def train_dataloader(self):
-        return DataLoader(self, batch_size=self.batch_size, num_workers=self.num_workers)
-
-    def val_dataloader(self):
-        dl = DataLoader(self, batch_size=self.batch_size, num_workers=self.num_workers)
-        if not self.fix_validation_set:
-            return dl
-        valid: List[Dict[str, Tensor]] = []
-        for _ in tqdm(range(self.valid_n_batches), desc="Generating fixed validation set"):
-            valid.append(next(iter(dl)))
-        return DataLoader(valid, batch_size=None, num_workers=0)
-
-    def test_dataloader(self):
-        return DataLoader(self, batch_size=self.batch_size, num_workers=self.num_workers)
-
-
-def _add_noise_and_background(image: Tensor, background: Tensor) -> Tensor:
-    image_with_background = image + background
-    noise = image_with_background.sqrt() * torch.randn_like(image_with_background)
-    return image_with_background + noise
-
-
-def _get_snr(image: Tensor, background: Tensor) -> float:
-    image_with_background = image + background
-    return torch.sqrt(torch.sum(image**2 / image_with_background)).item()
