+++ bliss/simulator/prior.py
@@ -1,6 +1,5 @@
 from typing import Optional
 
-import numpy as np
 import pytorch_lightning as pl
 import torch
 from einops import rearrange
@@ -103,7 +102,6 @@ class ImagePrior(pl.LightningModule):
         locs = self._sample_locs(is_on_array)
 
         galaxy_bools, star_bools = self._sample_n_galaxies_and_stars(is_on_array)
-        lensed_galaxy_bools = self._sample_n_lenses(is_on_array, galaxy_bools)
         galaxy_params = self._sample_galaxy_params(self.galaxy_prior, galaxy_bools)
         star_fluxes = self._sample_star_fluxes(star_bools)
         star_log_fluxes = self._get_log_fluxes(star_fluxes)
@@ -117,16 +115,6 @@ class ImagePrior(pl.LightningModule):
             "star_fluxes": star_fluxes,
             "star_log_fluxes": star_log_fluxes,
         }
-
-        if self.lensed_galaxy_prior is not None:
-            lensed_galaxy_params = self._sample_galaxy_params(
-                self.lensed_galaxy_prior, lensed_galaxy_bools
-            )
-            pure_lens_params = self._sample_lens_params(lensed_galaxy_bools)
-            lens_params = torch.cat((lensed_galaxy_params, pure_lens_params), dim=-1)
-
-            catalog_params["lensed_galaxy_bools"] = lensed_galaxy_bools
-            catalog_params["lens_params"] = lens_params
 
         return TileCatalog(tile_slen, catalog_params)
 
@@ -268,25 +256,3 @@ class ImagePrior(pl.LightningModule):
             n,
             device=device,
         )
-
-    def _sample_lens_params(self, lensed_galaxy_bools):
-        """Sample latent galaxy params from GalaxyPrior object."""
-        base_shape = list(lensed_galaxy_bools.shape)[:-1]
-        device = lensed_galaxy_bools.device
-        lens_params = self._sample_param_from_dist(base_shape, 5, torch.rand, device)
-        if self.prob_lensed_galaxy > 0.0:
-            # latents are: theta_E, center_x/y, e_1/2
-            base_radii = self._sample_param_from_dist(base_shape, 1, torch.rand, device)
-            base_centers = self._sample_param_from_dist(base_shape, 2, torch.randn, device)
-            base_qs = self._sample_param_from_dist(base_shape, 1, torch.rand, device)
-            base_betas = self._sample_param_from_dist(base_shape, 1, torch.rand, device)
-
-            lens_params[..., 0:1] = base_radii * 25.0 + 5.0
-            lens_params[..., 1:3] = base_centers * 1.0
-
-            # ellipticities must satisfy some angle relationships
-            beta_radians = (base_betas - 0.5) * (np.pi / 2)  # [-pi / 4, pi / 4]
-            ell_factors = (1 - base_qs) / (1 + base_qs)
-            lens_params[..., 3:4] = ell_factors * torch.cos(beta_radians)
-            lens_params[..., 4:5] = ell_factors * torch.sin(beta_radians)
-        return lens_params * lensed_galaxy_bools
