+++ case_studies/strong_lensing/lense_decoder.py
@@ -208,3 +208,25 @@ class LensedGalaxyTileDecoder(nn.Module):
         sized_galaxy = fit_source_to_ptile(galaxy, self.ptile_slen)
         outsize = sized_galaxy.shape[-1]
         return sized_galaxy.view(n_galaxies, self.n_bands, outsize, outsize)
+
+    def _sample_lens_params(self, lensed_galaxy_bools):
+        """Sample latent galaxy params from GalaxyPrior object."""
+        base_shape = list(lensed_galaxy_bools.shape)[:-1]
+        device = lensed_galaxy_bools.device
+        lens_params = self._sample_param_from_dist(base_shape, 5, torch.rand, device)
+        if self.prob_lensed_galaxy > 0.0:
+            # latents are: theta_E, center_x/y, e_1/2
+            base_radii = self._sample_param_from_dist(base_shape, 1, torch.rand, device)
+            base_centers = self._sample_param_from_dist(base_shape, 2, torch.randn, device)
+            base_qs = self._sample_param_from_dist(base_shape, 1, torch.rand, device)
+            base_betas = self._sample_param_from_dist(base_shape, 1, torch.rand, device)
+
+            lens_params[..., 0:1] = base_radii * 25.0 + 5.0
+            lens_params[..., 1:3] = base_centers * 1.0
+
+            # ellipticities must satisfy some angle relationships
+            beta_radians = (base_betas - 0.5) * (np.pi / 2)  # [-pi / 4, pi / 4]
+            ell_factors = (1 - base_qs) / (1 + base_qs)
+            lens_params[..., 3:4] = ell_factors * torch.cos(beta_radians)
+            lens_params[..., 4:5] = ell_factors * torch.sin(beta_radians)
+        return lens_params * lensed_galaxy_bools
